+++ src/apps/front/src/features/imagination/components/imagination.scss
@@ -63,8 +63,8 @@ textarea::placeholder {
 }
 
 .loading-img {
-  width: 512px;
-  height: 512px;
+  width: 500px;
+  height: 500px;
 }
 
 .settings-options {
@@ -95,10 +95,10 @@ textarea::placeholder {
   align-items: center;
   justify-content: flex-end;
   width: 100%;
-  .img-Generate img {
-    max-width: 100%;
-    max-height: 100%;
-  }
+}
+.img-Generate img {
+  max-width: 100%;
+  max-height: 100%;
 }
 
 .generate-btn-active {
